+++ src/app/blog/blog.module.ts
@@ -2,11 +2,11 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { PostComponent } from './post/post.component';
 import { AllComponent } from './all/all.component';
-import { CardComponent } from './all/card/card.component';
 import {BlogRoutingModule} from './blog-routing.module';
 import {MarkdownModule} from 'ngx-markdown';
 import { MatChipsModule } from '@angular/material/chips';
 import { CoreModule } from '../core/core.module';
+import { NgAisModule } from 'angular-instantsearch';
 
 @NgModule({
   imports: [
@@ -14,8 +14,9 @@ import { CoreModule } from '../core/core.module';
     BlogRoutingModule,
     MarkdownModule,
     MatChipsModule,
-    CoreModule
+    CoreModule,
+    NgAisModule
   ],
-  declarations: [PostComponent, AllComponent, CardComponent]
+  declarations: [PostComponent, AllComponent]
 })
 export class BlogModule { }